Add retries to service tomcat. Resolves Cook-1599
On el6 and some other systems, the init script to start tomcat returns
immediately, but the service can take several seconds to initialize.  When the
restart from writing configuration files comes, the tomcat service will not
respond to it properly. It returns an error that causes the recipe to fail.
The retry delay is set longer than we need for an empty tomcat server, but is
intended to avoid problems when you have a tomcat app installed that takes
longer to start.