Removes hard coded path to the keytool in the java home path
Adds an include to the default java recipe attributes and then modifies the default node['tomcat']['keytool'] to be dynamically populated with the keytool path rather than hard coded for each platform. I've found the Ubuntu path was using an incorrect alias to the java home path causing my own recipes to fail.