Fixed: test-kitchen on CentOS 6.6
There was a NoClassDefFoundError with a ModelMBeanNotificationBroadcaster on tomcat startup, which was really weird.
It turns out this was happening because the JAVA_HOME was not being set correctly and the GNU libgcj was getting picked up...
```
[root@tomcat7-centos-66 logs]# java -version
java version "1.5.0"
gij (GNU libgcj) version 4.4.7 20120313 (Red Hat 4.4.7-16)

Copyright (C) 2007 Free Software Foundation, Inc.
This is free software; see the source for copying conditions.  There is NO
warranty; not even for MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
```
Setting the set_etc_environment in test-kitchen on this platform resolved the issue.