Excluded centos 6.6 from tomcat7 test suite
It seems there is an issue with the java cookbook on this platform that does not allow the tomcat service to start.
For some reason, when test kitchen runs the java cookbook default recipe, before running the tomcat default recipe, if you login to the instance, java -version still doesn't give you the version of java you installed as the default.
Some hints: if i run kitchen converge again (i.e. yes, a second time), java version is fixed. So if i converge the instance twice, it runs normally.