Fixes #236
I believe this was probably the intended behaviour anyway (i.e. I'm not sure if the extra `/` was intentional, looking at the other exclude patterns). Keeping an empty ROOT dir doesn't really offer much anyway (unless not being able to deploy the ROOT app is a feature, for some reason?)