1st part of supporting non-system default init systems with proper restarts
You can use the tomcat_service_WHATEVER names currently, but we're not restarting with env vals change and we can't because we don't specify the init system to use within the provider. Basically the whole thing is currently broken. This isn't a complete fix since the sys-v script is a bit busted right now.